requirements.txt
Django>=4.2
pandas>=2.0
scikit-learn>=1.2
openpyxl>=3.1
Pillow>=10.0
spacy>=3.8
gunicorn>=21.2
whitenoise>=6.6
dj-database-url>=1.0
python-dotenv>=1.0